+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Undeleased
+# Version 2.1.0 (2026-05-21)
 
 ## New Features